+++ src/libcw/libcw_gen.h
@@ -200,7 +200,7 @@ struct cw_gen_struct {
 	/* Used to calculate sine wave.
 	   Phase offset needs to be stored between consecutive calls to
 	   function calculating consecutive fragments of sine wave. */
-	double phase_offset; /* TODO: change to float. */
+	float phase_offset;
